moves Contact and AddressBook in separate files
adds a basic application to browse contacts from ZPUI
