+++ apps/personal/contacts/vcard_converter.py
@@ -1,0 +1,53 @@
+import logging
+
+import vobject
+
+from apps.personal.contacts.address_book import Contact
+
+
+class VCardContactConverter(object):
+    vcard_mapping = {
+        'fn': 'name',
+        'n': 'name',
+        'adr': 'address',
+        'tel': 'telephone',
+        'email': 'email',
+        'url': 'url',
+        'note': 'note',
+        'org': 'org',
+        'title': 'title'
+    }
+
+    @staticmethod
+    def to_zpui_contact(v_contact):
+        # type: (vobject.base.Component) -> Contact
+        contact = Contact()
+        for key in v_contact.contents.keys():
+            if key not in VCardContactConverter.vcard_mapping: continue
+            attr = getattr(contact, VCardContactConverter.vcard_mapping[key])
+            assert type(attr) == list
+            attr += [v.value for v in v_contact.contents[key]]
+        return contact
+
+    @staticmethod
+    def parse_vcard_file(file_path):
+        # type: (str) -> list
+        file_content = VCardContactConverter.read_file_content(file_path)
+        contacts = [c for c in vobject.readComponents(file_content, ignoreUnreadable=True)]
+        logging.info("Found {} contact in file {}", len(contacts), file_path)
+        return contacts
+
+    @staticmethod
+    def read_file_content(file_path):
+        with open(file_path, 'r') as f:
+            file_content = f.read()
+            return file_content
+
+    @staticmethod
+    def from_vcards(contact_card_files):
+        # type: (list) -> list
+        contacts = []
+        for file_path in contact_card_files:
+            contacts += VCardContactConverter.parse_vcard_file(file_path)
+        logging.info("finished : {} contacts loaded", len(contacts))
+        return [VCardContactConverter.to_zpui_contact(c) for c in contacts]
